Develop

+++ lib/cms/model/rel/bracket.rb
@@ -14,17 +14,18 @@ module Cms::Model::Rel::Bracket
   private
 
   def body_with_bracket
-    if self.class.is_a?(Cms::Layout)
-      "#{body}#{mobile_body}#{smart_phone_body}"
-    else
-      body
-    end
+    text_columns = self.class.columns.select { |c| c.type == :text || c.type == :string }
+    text_columns.map { |c| send(c.name) }.join
   end
 
   def save_brackets
+    _site_id = respond_to?(:site_id) ? site_id : respond_to?(:content) && content ? content.site_id : nil
+    _concept_id = respond_to?(:concept_id) ? concept_id : nil
+    return unless _site_id
+
     brackets.destroy_all
     bracket_names.each do |name|
-      brackets.create(site_id: site_id, concept_id: concept_id, name: name)
+      brackets.create(site_id: _site_id, concept_id: _concept_id, name: name)
     end
   end
 end
